api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
yt-dlp>=2024.12.0
mutagen>=1.47.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pyjwt>=2.9.0
httpx>=0.27.0
mood-detector @ git+https://github.com/usr-wwelsh/mood-detector.git
